Backport first-part bugfixes from vesuvius-c (post-c5cb6cc)
Sync the public first-part pipeline (cube_mesh / grid_pipeline / grid_weld)
with fixes that landed in the private vesuvius-c tree after the last sync
point (vesuvius c5cb6cc == this repo's prior HEAD). Excludes all
ribbon/flatten/snap/unroll/whole work and the other new post-sync modules
(sheet_reweld, patch_repair, isotropic_remesh) and the QemOpts/prob_quadrics
decimation refactor.

Winding preservation (weld+decimate no longer break Hoppe orientation):
- qem.c: fix the maintenance edge-flip's reversed write -- was
  (c,d,a)+(d,c,b), now orientation-preserving (c,a,d)+(d,b,c); the angle/
  normal guard now matches the write. (vesuvius f944f68)
- weld_cleanup.c: the same reversed-write edge-flip fix. (5e0c916)
- vert_weld: Weld_verts gains guard_orient -- refuse to fuse a coincident
  recto/verso pair (a non-orientable knot no winding BFS can fix). grid_weld's
  micro-weld passes true; the point-dedup callers pass false. (f944f68)
- grid_weld: post-fill winding repair -- re-run OrientMesh_consistent after
  the late closers so fills wound to their own Newell normal are reconciled
  (topology-preserving; component count unchanged). (6793779)

Robustness / correctness:
- arena: widen Arena_alloc/Arena_calloc to size_t (fix >2 GB truncation on
  LLP64/Win64) plus 0-byte and calloc-overflow guards. (317f92f)
- pred_reject: reject "empty garbage" cubes whose largest 6-connected
  component is below MIN_CC_SIZE (they mesh to a guaranteed empty FAIL);
  grid_pipeline reports it as a distinct reject kind. (916020f)
- ball_pivot winding gate: also gate AXIAL chords (z-seam divot "connects
  up" fix) plus a final per-face cross-wrap filter. (dbe04fb)
- pinhole_fill: orientation-gated wide-bowtie close + 3-loop diameter
  exemption (PINHOLE_TRI_DIAM_VOX). (4be63ab, 6793779)

Seam-hole fill (missing-triangle dead zone):
- New self-contained module remesh/seam_hole_fill.{c,h}: merger-safe ear-clip
  of the small straddling seam punctures the BPA bridge leaves open, gated by
  near-seam+straddle / small-extent / winding-phase. (29a1b44)
- grid_weld: run it before the final manifold guard, plus a bounded
  post-guard convergence loop that re-runs the merger-safe closers after the
  guard opens notches. (29a1b44, 58b04e2)
- Wired into grid_weld.vcxproj and src/Makefile (WELD_OBJ).

Verified: scrollfiesta.sln Release|x64 builds clean (0 errors); end-to-end
grid_pipeline on real PHerc0139 cubes welds to a clean 2-manifold
(same_dir_pairs=0, non_manifold=0, pinch_verts=0); the seam-hole pass detects
and correctly gates seam candidates on a 2x2x2 z-seam block with no regression.

+++ src/common/arena.h
@@ -1,6 +1,8 @@
 #ifndef ARENA_INCLUDED
 #define ARENA_INCLUDED
 
+#include <stddef.h>
+
 #include "except.h"
 
 typedef struct Arena_T *Arena_T;
@@ -19,21 +21,27 @@ Arena_T Arena_new(void);
 void    Arena_dispose(Arena_T *ap);
 void    Arena_free(Arena_T arena);
 
-/* Allocation — never returns NULL (raises Arena_Failed on OOM) */
-void   *Arena_alloc(Arena_T arena, long nbytes, const char *file, int line);
-void   *Arena_calloc(Arena_T arena, long count, long size,
+/* Allocation — never returns NULL (raises Arena_Failed on OOM).
+ * Sizes are size_t (not long): a single allocation may exceed 2 GB on LLP64
+ * (Win64), where `long` is 32-bit and would silently truncate. */
+void   *Arena_alloc(Arena_T arena, size_t nbytes, const char *file, int line);
+void   *Arena_calloc(Arena_T arena, size_t count, size_t size,
                      const char *file, int line);
 
 /* Save/restore */
 Arena_Mark Arena_save(Arena_T arena);
 void       Arena_restore(Arena_T arena, Arena_Mark mark);
 
-/* Convenience macros — inject __FILE__, __LINE__ for diagnostics */
+/* Convenience macros — inject __FILE__, __LINE__ for diagnostics.
+ * The explicit (size_t) casts keep the ~1000 existing call sites (many of which
+ * still pass a legacy `(long)(...)` size) warning-clean under gcc -Wconversion
+ * without editing them; new code passing a plain size_t gets the full >2 GB
+ * range. Do NOT wrap the size in (long) in new code — that would re-truncate. */
 #define ARENA_ALLOC(a, nbytes) \
-    Arena_alloc((a), (nbytes), __FILE__, __LINE__)
+    Arena_alloc((a), (size_t)(nbytes), __FILE__, __LINE__)
 #define ARENA_CALLOC(a, count, size) \
-    Arena_calloc((a), (count), (size), __FILE__, __LINE__)
+    Arena_calloc((a), (size_t)(count), (size_t)(size), __FILE__, __LINE__)
 #define ARENA_NEW(a, p) \
-    ((p) = Arena_alloc((a), (long)sizeof *(p), __FILE__, __LINE__))
+    ((p) = Arena_alloc((a), sizeof *(p), __FILE__, __LINE__))
 
 #endif
